refactor: move CLI logic to cli/commands.py, __init__.py re-exports only
